CMD of star cluster M103
To devise an Imaging and spectroscopic classification of the stars in the selected star cluster M103. First part of lab was aimed at drawing the color-magnitude diagram of the said star cluster by doing an analysis on the raw images obtained using the TOBI telescope. Later on we took spectropic data of the same cluster on one of the stars of the cluster , whereby we plotted its spectrum and tried to compute various properties of the star like its surface temperature and Redshift.